+++ packages/std/src/exports.rs
@@ -23,7 +23,7 @@ use crate::ibc::{
 };
 use crate::ibc::{IbcChannelOpenMsg, IbcChannelOpenResponse};
 #[cfg(feature = "ibc2")]
-use crate::ibc2::Ibc2PacketReceiveMsg;
+use crate::ibc2::{Ibc2PacketReceiveMsg, Ibc2PacketTimeoutMsg};
 use crate::imports::{ExternalApi, ExternalQuerier, ExternalStorage};
 use crate::memory::{Owned, Region};
 use crate::panic::install_panic_handler;
@@ -455,8 +455,7 @@ where
 /// do_ibc_packet_timeout is designed for use with #[entry_point] to make a "C" extern
 ///
 /// contract_fn is called when a packet that this contract previously sent has provably
-/// timedout and will never be relayed to the calling chain. This generally behaves
-/// like ick_ack_fn upon an acknowledgement containing an error.
+/// timed out and will never be relayed to the destination chain.
 ///
 /// - `Q`: custom query type (see QueryRequest)
 /// - `C`: custom response message type (see CosmosMsg)
@@ -547,6 +546,35 @@ where
 {
     install_panic_handler();
     let res = _do_ibc2_packet_receive(
+        contract_fn,
+        env_ptr as *mut Region<Owned>,
+        msg_ptr as *mut Region<Owned>,
+    );
+    let v = to_json_vec(&res).unwrap();
+    Region::from_vec(v).to_heap_ptr() as u32
+}
+
+/// do_ibc2_packet_timeout is designed for use with #[entry_point] to make a "C" extern
+///
+/// contract_fn is called when a packet that this contract previously sent has provably
+/// timed out and will never be relayed to the destination chain.
+///
+/// - `Q`: custom query type (see QueryRequest)
+/// - `C`: custom response message type (see CosmosMsg)
+/// - `E`: error type for responses
+#[cfg(feature = "ibc2")]
+pub fn do_ibc2_packet_timeout<Q, C, E>(
+    contract_fn: &dyn Fn(DepsMut<Q>, Env, Ibc2PacketTimeoutMsg) -> Result<IbcBasicResponse<C>, E>,
+    env_ptr: u32,
+    msg_ptr: u32,
+) -> u32
+where
+    Q: CustomQuery,
+    C: CustomMsg,
+    E: ToString,
+{
+    install_panic_handler();
+    let res = _do_ibc2_packet_timeout(
         contract_fn,
         env_ptr as *mut Region<Owned>,
         msg_ptr as *mut Region<Owned>,
@@ -941,6 +969,29 @@ where
 
     let env: Env = try_into_contract_result!(from_json(env));
     let msg: Ibc2PacketReceiveMsg = try_into_contract_result!(from_json(msg));
+
+    let mut deps = make_dependencies();
+    contract_fn(deps.as_mut(), env, msg).into()
+}
+
+#[cfg(feature = "ibc2")]
+fn _do_ibc2_packet_timeout<Q, C, E>(
+    contract_fn: &dyn Fn(DepsMut<Q>, Env, Ibc2PacketTimeoutMsg) -> Result<IbcBasicResponse<C>, E>,
+    env_ptr: *mut Region<Owned>,
+    msg_ptr: *mut Region<Owned>,
+) -> ContractResult<IbcBasicResponse<C>>
+where
+    Q: CustomQuery,
+    C: CustomMsg,
+    E: ToString,
+{
+    let env: Vec<u8> =
+        unsafe { Region::from_heap_ptr(ptr::NonNull::new(env_ptr).unwrap()).into_vec() };
+    let msg: Vec<u8> =
+        unsafe { Region::from_heap_ptr(ptr::NonNull::new(msg_ptr).unwrap()).into_vec() };
+
+    let env: Env = try_into_contract_result!(from_json(env));
+    let msg: Ibc2PacketTimeoutMsg = try_into_contract_result!(from_json(msg));
 
     let mut deps = make_dependencies();
     contract_fn(deps.as_mut(), env, msg).into()