+++ packages/std/src/lib.rs
@@ -77,7 +77,7 @@ pub use crate::ibc::{
     IbcSourceCallbackMsg, IbcSrcCallback, IbcTimeout, IbcTimeoutBlock, IbcTimeoutCallbackMsg,
     TransferMsgBuilder,
 };
-pub use crate::ibc2::{Ibc2Msg, Ibc2PacketReceiveMsg, Ibc2Payload};
+pub use crate::ibc2::{Ibc2Msg, Ibc2PacketReceiveMsg, Ibc2PacketTimeoutMsg, Ibc2Payload};
 #[cfg(feature = "iterator")]
 pub use crate::iterator::{Order, Record};
 pub use crate::math::{
@@ -90,14 +90,14 @@ pub use crate::msgpack::{from_msgpack, to_msgpack_binary, to_msgpack_vec};
 pub use crate::never::Never;
 pub use crate::pagination::PageRequest;
 pub use crate::query::{
-    AllBalanceResponse, AllDelegationsResponse, AllDenomMetadataResponse, AllValidatorsResponse,
-    BalanceResponse, BankQuery, BondedDenomResponse, ChannelResponse, CodeInfoResponse,
-    ContractInfoResponse, CustomQuery, DecCoin, Delegation, DelegationResponse,
-    DelegationRewardsResponse, DelegationTotalRewardsResponse, DelegatorReward,
-    DelegatorValidatorsResponse, DelegatorWithdrawAddressResponse, DenomMetadataResponse,
-    DistributionQuery, FeeEnabledChannelResponse, FullDelegation, GrpcQuery, IbcQuery,
-    ListChannelsResponse, PortIdResponse, QueryRequest, StakingQuery, SupplyResponse, Validator,
-    ValidatorResponse, WasmQuery,
+    AllDelegationsResponse, AllDenomMetadataResponse, AllValidatorsResponse, BalanceResponse,
+    BankQuery, BondedDenomResponse, ChannelResponse, CodeInfoResponse, ContractInfoResponse,
+    CustomQuery, DecCoin, Delegation, DelegationResponse, DelegationRewardsResponse,
+    DelegationTotalRewardsResponse, DelegatorReward, DelegatorValidatorsResponse,
+    DelegatorWithdrawAddressResponse, DenomMetadataResponse, DistributionQuery,
+    FeeEnabledChannelResponse, FullDelegation, GrpcQuery, IbcQuery, ListChannelsResponse,
+    PortIdResponse, QueryRequest, StakingQuery, SupplyResponse, Validator, ValidatorResponse,
+    WasmQuery,
 };
 #[cfg(all(feature = "stargate", feature = "cosmwasm_1_2"))]
 pub use crate::results::WeightedVoteOption;
@@ -124,8 +124,6 @@ mod imports;
 #[cfg(target_arch = "wasm32")]
 mod memory; // Used by exports and imports only. This assumes pointers are 32 bit long, which makes it untestable on dev machines.
 
-#[cfg(all(feature = "ibc2", target_arch = "wasm32"))]
-pub use crate::exports::do_ibc2_packet_receive;
 #[cfg(all(feature = "cosmwasm_2_2", target_arch = "wasm32"))]
 pub use crate::exports::do_migrate_with_info;
 #[cfg(target_arch = "wasm32")]
@@ -133,6 +131,8 @@ pub use crate::exports::{
     do_execute, do_ibc_destination_callback, do_ibc_source_callback, do_instantiate, do_migrate,
     do_query, do_reply, do_sudo,
 };
+#[cfg(all(feature = "ibc2", target_arch = "wasm32"))]
+pub use crate::exports::{do_ibc2_packet_receive, do_ibc2_packet_timeout};
 #[cfg(all(feature = "stargate", target_arch = "wasm32"))]
 pub use crate::exports::{
     do_ibc_channel_close, do_ibc_channel_connect, do_ibc_channel_open, do_ibc_packet_ack,