+++ packages/vm/src/cache.rs
@@ -340,7 +340,9 @@ where
             has_ibc_entry_points: REQUIRED_IBC_EXPORTS
                 .iter()
                 .all(|required| exports.contains(required.as_ref())),
-            has_ibc2_entry_points: exports.contains(REQUIRED_IBC2_EXPORT.as_ref()),
+            has_ibc2_entry_points: REQUIRED_IBC2_EXPORT
+                .iter()
+                .all(|required| exports.contains(required.as_ref())),
             entrypoints,
             required_capabilities: required_capabilities_from_module(&module)
                 .into_iter()
@@ -611,7 +613,6 @@ fn remove_wasm_from_disk(dir: impl Into<PathBuf>, checksum: &Checksum) -> VmResu
 mod tests {
     use super::*;
     use crate::calls::{call_execute, call_instantiate};
-    use crate::capabilities::capabilities_from_csv;
     use crate::testing::{mock_backend, mock_env, mock_info, MockApi, MockQuerier, MockStorage};
     use cosmwasm_std::{coins, Empty};
     use std::borrow::Cow;
@@ -640,7 +641,19 @@ mod tests {
     "#;
 
     fn default_capabilities() -> HashSet<String> {
-        capabilities_from_csv("iterator,staking")
+        HashSet::from([
+            "cosmwasm_1_1".to_string(),
+            "cosmwasm_1_2".to_string(),
+            "cosmwasm_1_3".to_string(),
+            "cosmwasm_1_4".to_string(),
+            "cosmwasm_1_4".to_string(),
+            "cosmwasm_2_0".to_string(),
+            "cosmwasm_2_1".to_string(),
+            "cosmwasm_2_2".to_string(),
+            "iterator".to_string(),
+            "staking".to_string(),
+            "stargate".to_string(),
+        ])
     }
 
     fn make_testing_options() -> CacheOptions {
@@ -693,7 +706,7 @@ mod tests {
 
         // execute
         let info = mock_info(&verifier, &coins(15, "earth"));
-        let msg = br#"{"release":{}}"#;
+        let msg = br#"{"release":{"denom":"earth"}}"#;
         let response = call_execute::<_, _, _, Empty>(instance, &mock_env(), &info, msg)
             .unwrap()
             .unwrap();
@@ -1137,7 +1150,7 @@ mod tests {
 
             // execute
             let info = mock_info(&verifier, &coins(15, "earth"));
-            let msg = br#"{"release":{}}"#;
+            let msg = br#"{"release":{"denom":"earth"}}"#;
             let response = call_execute::<_, _, _, Empty>(&mut instance, &mock_env(), &info, msg)
                 .unwrap()
                 .unwrap();
@@ -1171,7 +1184,7 @@ mod tests {
 
             // execute
             let info = mock_info(&verifier, &coins(15, "earth"));
-            let msg = br#"{"release":{}}"#;
+            let msg = br#"{"release":{"denom":"earth"}}"#;
             let response = call_execute::<_, _, _, Empty>(&mut instance, &mock_env(), &info, msg)
                 .unwrap()
                 .unwrap();
@@ -1207,7 +1220,7 @@ mod tests {
 
             // execute
             let info = mock_info(&verifier, &coins(15, "earth"));
-            let msg = br#"{"release":{}}"#;
+            let msg = br#"{"release":{"denom":"earth"}}"#;
             let response = call_execute::<_, _, _, Empty>(&mut instance, &mock_env(), &info, msg)
                 .unwrap()
                 .unwrap();
@@ -1280,7 +1293,7 @@ mod tests {
             .get_instance(&checksum, backend2, TESTING_OPTIONS)
             .unwrap();
         let info = mock_info(&bob, &coins(15, "earth"));
-        let msg = br#"{"release":{}}"#;
+        let msg = br#"{"release":{"denom":"earth"}}"#;
         let res = call_execute::<_, _, _, Empty>(&mut instance, &mock_env(), &info, msg).unwrap();
         let msgs = res.unwrap().messages;
         assert_eq!(1, msgs.len());
@@ -1290,7 +1303,7 @@ mod tests {
             .get_instance(&checksum, backend1, TESTING_OPTIONS)
             .unwrap();
         let info = mock_info(&sue, &coins(15, "earth"));
-        let msg = br#"{"release":{}}"#;
+        let msg = br#"{"release":{"denom":"earth"}}"#;
         let res = call_execute::<_, _, _, Empty>(&mut instance, &mock_env(), &info, msg).unwrap();
         let msgs = res.unwrap().messages;
         assert_eq!(1, msgs.len());
@@ -1468,15 +1481,24 @@ mod tests {
                     E::Execute,
                     E::Query
                 ]),
-                required_capabilities: BTreeSet::new(),
-                contract_migrate_version: Some(42),
+                required_capabilities: BTreeSet::from([
+                    "cosmwasm_1_1".to_string(),
+                    "cosmwasm_1_2".to_string(),
+                    "cosmwasm_1_3".to_string(),
+                    "cosmwasm_1_4".to_string(),
+                    "cosmwasm_1_4".to_string(),
+                    "cosmwasm_2_0".to_string(),
+                    "cosmwasm_2_1".to_string(),
+                    "cosmwasm_2_2".to_string(),
+                ]),
+                contract_migrate_version: Some(420),
             }
         );
 
         let checksum2 = cache.store_code(IBC_REFLECT, true, true).unwrap();
         let report2 = cache.analyze(&checksum2).unwrap();
         let mut ibc_contract_entrypoints =
-            BTreeSet::from([E::Instantiate, E::Migrate, E::Reply, E::Query]);
+            BTreeSet::from([E::Instantiate, E::Migrate, E::Execute, E::Reply, E::Query]);
         ibc_contract_entrypoints.extend(REQUIRED_IBC_EXPORTS);
         assert_eq!(
             report2,
@@ -1485,6 +1507,14 @@ mod tests {
                 has_ibc2_entry_points: false,
                 entrypoints: ibc_contract_entrypoints,
                 required_capabilities: BTreeSet::from_iter([
+                    "cosmwasm_1_1".to_string(),
+                    "cosmwasm_1_2".to_string(),
+                    "cosmwasm_1_3".to_string(),
+                    "cosmwasm_1_4".to_string(),
+                    "cosmwasm_1_4".to_string(),
+                    "cosmwasm_2_0".to_string(),
+                    "cosmwasm_2_1".to_string(),
+                    "cosmwasm_2_2".to_string(),
                     "iterator".to_string(),
                     "stargate".to_string()
                 ]),
@@ -1530,7 +1560,7 @@ mod tests {
         let checksum5 = cache.store_code(IBC2, true, true).unwrap();
         let report5 = cache.analyze(&checksum5).unwrap();
         let mut ibc2_contract_entrypoints = BTreeSet::from([E::Instantiate, E::Query]);
-        ibc2_contract_entrypoints.extend([REQUIRED_IBC2_EXPORT]);
+        ibc2_contract_entrypoints.extend(REQUIRED_IBC2_EXPORT);
         assert_eq!(
             report5,
             AnalysisReport {