+++ CHANGELOG.md
@@ -15,14 +15,15 @@ and this project adheres to
   `From<Int128> for i128` ([#2268])
 - cosmwasm-std: Implement `Uint128::from_{be,le}_bytes` and
   `Uint64::from_{be,le}_bytes`. ([#2269])
-- cosmwasm-std: Added new `Ibc2Msg` and `CosmosMsg::Ibc2` variant ([#2390],
-  [#2403])
-- cosmwasm-std: Implement downcasting for `Api` trait. This allows using
-  `MockApi::addr_make` from `DepsMut`. ([#2383])
+- cosmwasm-std: Added new `Ibc2Msg` and `CosmosMsg::Ibc2` variant ([#2340],
+  [#2390], [#2403])
 - cosmwasm-std: Added `ibc2_port` to `ContractInfoResponse`. ([#2390], [#2403])
 - cosmwasm-vm: Added `ibc2_packet_receive` entrypoint ([#2403])
 - cosmwasm-vm: Add IBC Callbacks entrypoints to the `Entrypoints` enum.
   ([#2438])
+- cosmwasm-std: Add `WriteAcknowledgement` to `Ibc2Msg` - ([#2425])
+- cosmwasm-vm: Add `ibc2_packet_timeout` entrypoint - ([#2454])
+- cosmwasm-std: Add `Ibc2PacketTimeoutMsg` message - ([#2454])
 
 ## Changed
 
@@ -60,16 +61,22 @@ and this project adheres to
 - cosmwasm-std: Remove previously deprecated `cosmwast_std::testing::mock_info`.
   Use `cosmwasm_std::testing::message_info` instead. ([#2393])
 - cosmwasm-std: Remove abort feature. ([#2141])
+- cosmwasm-std: Change `Coin::amount` to `Uint256` instead of `Uint128`.
+  ([#2458])
 - cosmwasm-std: Replace dependency `serde-json-wasm` with `serde_json`.
   ([#2195])
 - cosmwasm-std: Make `GovMsg` `#[non_exhaustive]` for consistency. ([#2347])
 - cosmwasm-crypto: Upgrade ark-\* dependencies to 0.5.0. ([#2432])
+- cosmwasm-std: Remove support for `BankQuery::AllBalances` and
+  `query_all_balances`. ([#2433])
+- cosmwasm-std: source_client instead of channel_id in IBCv2 - ([#2450])
 
 ## Fixed
 
 - cosmwasm-schema: The schema export now doesn't overwrite existing
   `additionalProperties` values anymore ([#2310])
 - cosmwasm-vm: Fix CWA-2025-002.
+- cosmwasm-std: Fix deserialization of `DenomMetadata`. ([#2417])
 
 [#2141]: https://github.com/CosmWasm/cosmwasm/issues/2141
 [#2155]: https://github.com/CosmWasm/cosmwasm/issues/2155
@@ -88,13 +95,18 @@ and this project adheres to
 [#2367]: https://github.com/CosmWasm/cosmwasm/issues/2367
 [#2374]: https://github.com/CosmWasm/cosmwasm/issues/2374
 [#2378]: https://github.com/CosmWasm/cosmwasm/issues/2378
-[#2383]: https://github.com/CosmWasm/cosmwasm/issues/2383
 [#2390]: https://github.com/CosmWasm/cosmwasm/issues/2390
 [#2393]: https://github.com/CosmWasm/cosmwasm/issues/2393
 [#2399]: https://github.com/CosmWasm/cosmwasm/pull/2399
 [#2403]: https://github.com/CosmWasm/cosmwasm/pull/2403
+[#2417]: https://github.com/CosmWasm/cosmwasm/pull/2417
+[#2425]: https://github.com/CosmWasm/cosmwasm/pull/2425
 [#2432]: https://github.com/CosmWasm/cosmwasm/pull/2432
+[#2433]: https://github.com/CosmWasm/cosmwasm/pull/2433
 [#2438]: https://github.com/CosmWasm/cosmwasm/pull/2438
+[#2450]: https://github.com/CosmWasm/cosmwasm/pull/2450
+[#2454]: https://github.com/CosmWasm/cosmwasm/pull/2454
+[#2458]: https://github.com/CosmWasm/cosmwasm/pull/2458
 
 ## [2.2.0] - 2024-12-17